refactor group functions in 3 modules: "geometry, analysis, visualization"
